[Refactor] 사용자 정보 템플릿 스타일 파일 분리
